+++ src/main/java/com/ecommerce/service/impl/UserServiceImpl.java
@@ -22,7 +22,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUserById(Long id) {
+    public User getUserById(int id) {
         return userDao.getUserById(id);
     }
 
@@ -38,7 +38,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public boolean deleteUser(Long id) {
+    public boolean deleteUser(int id) {
         User user = userDao.getUserById(id);
         if (user != null) {
             userDao.deleteUser(id);
